moving thr history elements to a separate plugin instead of index
