+++ docs/.vitepress/theme/index.ts
@@ -16,61 +16,9 @@ import BannerSponsor from './components/BannerSponsor.vue'
 import AsideSponsors from './components/AsideSponsors.vue'
 import Donate from './components/Donate.vue'
 import FooterLinks from './components/FooterLinks.vue';
-import variableCommands from "vitepress-md-var";
 import ExegolHistoryImport from './components/ExegolHistoryImport.vue'
+import thrHistory from '../plugins/thrHistory.ts'
 const isMobileorTablet = useMediaQuery('(max-width: 1279px)')
-
-
-// Storage key for persisting variables
-const STORAGE_KEY = 'thr_commands_variables';
-// 24 hours in milliseconds
-const EXPIRY_TIME = 24 * 60 * 60 * 1000;
-
-/**
- * Stores variable value to localStorage with expiration
- * @param varName - name of variable to store
- * @param varValue - value of variable to store
- */
-function storeVar(varName, varValue) {
-  const values = loadVarValues();
-  try {
-    values.set(varName, varValue);
-    const dataToStore = {
-      timestamp: Date.now(),
-      values: Object.fromEntries(values)
-    };
-    localStorage.setItem(STORAGE_KEY, JSON.stringify(dataToStore));
-  } catch (e) {
-    console.error('Failed to save variable values:', e);
-  }
-}
-
-/**
- * Loads variable values from localStorage if not expired
- * @returns Map of variable names and their saved values
- */
-function loadVarValues() {
-  try {
-    const saved = localStorage.getItem(STORAGE_KEY);
-    if (!saved) return new Map();
-    
-    const data = JSON.parse(saved);
-    const now = Date.now();
-    
-    // Check if data has expired (24 hours)
-    if (data.timestamp && now - data.timestamp > EXPIRY_TIME) {
-      // Data expired, remove it and return empty map
-      localStorage.removeItem(STORAGE_KEY);
-      return new Map();
-    }
-    
-    // Data still valid
-    return new Map(Object.entries(data.values || {}));
-  } catch (e) {
-    console.error('Failed to load variable values:', e);
-    return new Map();
-  }
-}
 
 export default {
   extends: DefaultTheme,
@@ -90,12 +38,15 @@ export default {
     })
   },
 
-  enhanceApp({ app }) {
+  enhanceApp({ app, router }) {
     app.component('VPSidebarItem', CustomSidebarItem);
     enhanceAppWithTabs(app);
     app.component('Donaters', Donaters)
     app.component('DonationPricingTable', DonationPricingTable)
     app.component('ExegolHistoryImport', ExegolHistoryImport)
+    
+    // Initialize thrHistory plugin with router
+    thrHistory(router.route);
   },
 
   // IMG ZOOM SETUP
@@ -115,11 +66,5 @@ export default {
       () => route.path,
       () => nextTick(() => initZoom())
     )
-
-    variableCommands(route, {
-      loadVar: varName => loadVarValues().get(varName),
-      storeVar: (varName, varValue) => storeVar(varName, varValue),
-      styling: "thr",
-    })
   },
 } satisfies Theme
